+++ Cargo.toml
@@ -40,6 +40,8 @@ terminal_size = "0.4.2"
 clap_complete = "4.5.50"
 walkdir = "2.5.0"
 once_cell = "1.21.3"
+strum = {version = "0.27.1", features = ["strum_macros"]}
+strum_macros = "0.27.1"
 
 
 [dev-dependencies]
