+++ src/projects.rs
@@ -540,6 +540,7 @@ pub async fn move_task_to_project(
 mod tests {
     use super::*;
     use crate::test;
+    use crate::test::responses::ResponseFromFile;
     use pretty_assertions::assert_eq;
 
     #[tokio::test]
@@ -562,7 +563,7 @@ mod tests {
             .mock("GET", "/api/v1/projects?limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::projects_response())
+            .with_body(ResponseFromFile::Projects.read().await)
             .create_async()
             .await;
 
@@ -589,7 +590,7 @@ mod tests {
             .mock("GET", "/api/v1/tasks/?project_id=123&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::today_tasks_response().await)
+            .with_body(ResponseFromFile::TodayTasks.read().await)
             .create_async()
             .await;
 
@@ -600,7 +601,7 @@ mod tests {
             )
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::comments_response())
+            .with_body(ResponseFromFile::CommentsAllTypes.read().await)
             .create_async()
             .await;
 
@@ -632,7 +633,7 @@ mod tests {
             .mock("GET", "/api/v1/projects?limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::new_projects_response())
+            .with_body(ResponseFromFile::NewProjects.read().await)
             .create_async()
             .await;
 
@@ -668,7 +669,7 @@ mod tests {
             .mock("GET", "/api/v1/projects?limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::new_projects_response())
+            .with_body(ResponseFromFile::NewProjects.read().await)
             .create_async()
             .await;
 
@@ -711,7 +712,7 @@ mod tests {
             .mock("GET", "/api/v1/tasks/?project_id=123&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::today_tasks_response().await)
+            .with_body(ResponseFromFile::TodayTasks.read().await)
             .create_async()
             .await;
 
@@ -719,7 +720,7 @@ mod tests {
             .mock("POST", "/api/v1/tasks/6Xqhv4cwxgjwG9w8/move")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::sync())
+            .with_body(ResponseFromFile::Task.read().await)
             .create_async()
             .await;
 
@@ -727,7 +728,7 @@ mod tests {
             .mock("GET", "/api/v1/sections?project_id=123&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::sections())
+            .with_body(ResponseFromFile::Sections.read().await)
             .create_async()
             .await;
 
@@ -735,7 +736,7 @@ mod tests {
             .mock("GET", "/api/v1/id_mappings/projects/123")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::ids())
+            .with_body(ResponseFromFile::Ids.read().await)
             .create_async()
             .await;
         let mock5 = server
@@ -745,7 +746,7 @@ mod tests {
             )
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::comments_response())
+            .with_body(ResponseFromFile::CommentsAllTypes.read().await)
             .create_async()
             .await;
 
@@ -786,7 +787,7 @@ mod tests {
             .mock("GET", "/api/v1/tasks/?project_id=123&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::today_tasks_response().await)
+            .with_body(ResponseFromFile::TodayTasks.read().await)
             .create_async()
             .await;
 
@@ -808,7 +809,7 @@ mod tests {
             .mock("DELETE", "/api/v1/projects/123")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::projects())
+            .with_body(ResponseFromFile::Project.read().await)
             .create_async()
             .await;
 
@@ -833,7 +834,7 @@ mod tests {
             .mock("GET", "/api/v1/tasks/?project_id=123&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::unscheduled_tasks_response().await)
+            .with_body(ResponseFromFile::UnscheduledTasks.read().await)
             .create_async()
             .await;
 
@@ -841,7 +842,7 @@ mod tests {
             .mock("POST", "/rest/v2/tasks/999999")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::today_task().await)
+            .with_body(ResponseFromFile::TodayTask.read().await)
             .create_async()
             .await;
 
@@ -849,14 +850,14 @@ mod tests {
             .mock("GET", "/api/v1/id_mappings/projects/123")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::ids())
+            .with_body(ResponseFromFile::Ids.read().await)
             .create_async()
             .await;
         let mock4 = server
             .mock("GET", "/api/v1/comments/?task_id=999999&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::comments_response())
+            .with_body(ResponseFromFile::CommentsAllTypes.read().await)
             .create_async()
             .await;
         let config = test::fixtures::config()
@@ -912,7 +913,7 @@ mod tests {
             .mock("GET", "/api/v1/tasks/?project_id=123&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::unscheduled_tasks_response().await)
+            .with_body(ResponseFromFile::UnscheduledTasks.read().await)
             .create_async()
             .await;
 
@@ -920,7 +921,7 @@ mod tests {
             .mock("POST", "/rest/v2/tasks/999999")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::today_task().await)
+            .with_body(ResponseFromFile::TodayTask.read().await)
             .create_async()
             .await;
 
@@ -928,7 +929,7 @@ mod tests {
             .mock("GET", "/api/v1/id_mappings/projects/123")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::ids())
+            .with_body(ResponseFromFile::Ids.read().await)
             .create_async()
             .await;
         let config = test::fixtures::config()
@@ -941,7 +942,7 @@ mod tests {
             .mock("GET", "/api/v1/comments/?task_id=999999&limit=200")
             .with_status(200)
             .with_header("content-type", "application/json")
-            .with_body(test::responses::comments_response())
+            .with_body(ResponseFromFile::CommentsAllTypes.read().await)
             .create_async()
             .await;
         let binding = config.projects().await.unwrap();
