--- tests/fixtures/comments_all_types.json
@@ -1,157 +1,0 @@
-{
-    "results": [
-        {
-            "id": "file-1",
-            "content": "File upload",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:00:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": {
-                "file_name": "file.pdf",
-                "file_size": 123456,
-                "file_type": "application/pdf",
-                "file_url": "https://example.com/file.pdf",
-                "image": null,
-                "image_height": null,
-                "image_width": null,
-                "resource_type": "file",
-                "upload_state": "completed",
-                "tn_s": null,
-                "tn_m": null,
-                "tn_l": null
-            }
-        },
-        {
-            "id": "video-1",
-            "content": "Video link",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:01:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": {
-                "html": "<iframe src=\"https://example.com/embed/video\"></iframe>",
-                "image": "https://example.com/preview.jpg",
-                "image_height": 360,
-                "image_width": 480,
-                "resource_type": "video",
-                "site_name": "YouTube",
-                "title": "Test Video",
-                "url": "https://example.com/video",
-                "tn_s": null,
-                "tn_m": null,
-                "tn_l": null,
-                "upload_state": "completed"
-            }
-        },
-        {
-            "id": "image-1",
-            "content": "Image preview",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:02:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": {
-                "resource_type": "image",
-                "url": "https://example.com/image.jpg",
-                "image": "https://example.com/image.jpg",
-                "image_height": 600,
-                "image_width": 800,
-                "site_name": "ImageHost",
-                "title": "Example Image",
-                "tn_s": null,
-                "tn_m": null,
-                "tn_l": null
-            }
-        },
-        {
-            "id": "url-1",
-            "content": "Full URL",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:03:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": {
-                "resource_type": "url",
-                "description": "A useful article",
-                "favicon": "https://site.com/favicon.ico",
-                "image": "https://site.com/image.png",
-                "image_height": 100,
-                "image_width": 200,
-                "site_name": "SiteName",
-                "title": "Interesting Article",
-                "url": "https://site.com/article"
-            }
-        },
-        {
-            "id": "shorturl-1",
-            "content": "Short URL",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:04:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": {
-                "resource_type": "url_short",
-                "title": "Shortened Link",
-                "url": "https://bit.ly/xyz123"
-            }
-        },
-        {
-            "id": "rich-1",
-            "content": "Rich embed",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:05:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": {
-                "html": "<iframe src=\"https://nytimes.com/embed\"></iframe>",
-                "image": "https://nytimes.com/thumb.jpg",
-                "image_height": 400,
-                "image_width": 800,
-                "resource_type": "rich",
-                "site_name": "New York Times",
-                "title": "News Story",
-                "url": "https://nytimes.com/story",
-                "tn_s": null,
-                "tn_m": null,
-                "tn_l": null,
-                "upload_state": "completed"
-            }
-        },
-        {
-            "id": "noattach-1",
-            "content": "Just a plain comment",
-            "is_deleted": false,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:06:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": null
-        },
-        {
-            "id": "deleted-1",
-            "content": "Should be skipped",
-            "is_deleted": true,
-            "posted_uid": "123",
-            "uids_to_notify": null,
-            "posted_at": "2025-01-01T12:07:00Z",
-            "reactions": null,
-            "item_id": "task-1",
-            "file_attachment": null
-        }
-    ],
-    "next_cursor": null
-}
