+++ tests/responses/Comment.json
@@ -1,0 +1,13 @@
+{
+  "content": "Need one bottle of milk",
+  "id": "2992679862",
+  "posted_at": "2016-09-22T07:00:00.000000Z",
+  "item_id": "123",
+  "is_deleted": false,
+  "attachment": {
+    "file_name": "File.pdf",
+    "file_type": "application/pdf",
+    "file_url": "https://s3.amazonaws.com/domorebetter/Todoist+Setup+Guide.pdf",
+    "resource_type": "file"
+  }
+}
