+++ tests/responses/Label.json
@@ -1,0 +1,7 @@
+{
+  "id": "123",
+  "name": "345",
+  "is_favorite": false,
+  "order": null,
+  "color": "red"
+}
