+++ tests/responses/TodayTask.json
@@ -1,0 +1,40 @@
+{
+  "user_id": "910",
+  "id": "6Xqhv4cwxgjwG9w8",
+  "project_id": "6VRRxv8CM6GVmmgf",
+  "section_id": null,
+  "parent_id": null,
+  "added_by_uid": "633166",
+  "assigned_by_uid": null,
+  "responsible_uid": null,
+  "labels": [
+    "computer"
+  ],
+  "deadline": {
+    "date": "INSERTDATE",
+    "lang": "en"
+  },
+  "duration": {
+    "amount": 15,
+    "unit": "minute"
+  },
+  "checked": false,
+  "is_deleted": false,
+  "added_at": "INSERTDATET22:29:34.404051Z",
+  "completed_at": null,
+  "updated_at": "INSERTDATET22:32:46.415849Z",
+  "due": {
+    "date": "INSERTDATET12:00:00Z",
+    "timezone": "America/Vancouver",
+    "string": "INSERTDATE 15:00",
+    "lang": "en",
+    "is_recurring": false
+  },
+  "priority": 3,
+  "child_order": 1,
+  "content": "TEST",
+  "description": "",
+  "note_count": 0,
+  "day_order": -1,
+  "is_collapsed": false
+}
